Separated notification logic from Firebase Messaging Service for platform agnostic approach
